Update friendly type for display
Patches to HeapUtil to keep it from throwing NPE on allocations of
multidimensional arrays with some dimensions not specified yet.

Also unit tests to show this working. I am not positive the numbers
are coming out right, but it looks good so far.

Future todo - add more tests of arrays of all types.